fcntl locks, faccessat, libc work, misc
added __stack_chk_fail_local to Scrt1.s
added memchr, memmem
optimized memcmp, strlen
fixed signatures of strcmp, strncmp, memcmp
added strcat, strspn, strstr
fixed utimes calling unlink :3
fixed some wrong printf format specifiers

added thread_queue_add_nonreentrant and used it to fix pipe race
added euid and egid fields to process_t
optimized sys_sigsuspend

+++ src/kernel.c
@@ -245,7 +245,7 @@ void kernel_print_cpu_info() {
 extern struct tss_segment tss;
 extern struct idt_gate * idt_descriptor_entries;
 
-uintptr_t boot_mem_top = 0; // the top of taken memory (either &_kernel_top, or the highest multiboot module)
+unsigned long boot_mem_top = 0; // the top of taken memory (either &_kernel_top, or the highest multiboot module)
 void * kernel_mem_top = NULL; // the top of all kernel memory (identity map, heap, page frame table)
 
 time_t system_time_sec = 0;
